Conflicts:
	Bundles/Raven.Bundles.Tests/App.config
	Imports/Newtonsoft.Json/Src/Newtonsoft.Json/Utilities/TypeExtensions.cs
	Raven.Client.Lightweight/Changes/RemoteDatabaseChanges.cs
	Raven.Client.Lightweight/Connection/Async/AsyncServerClient.cs
	Raven.Client.Lightweight/Connection/IDatabaseCommands.cs
	Raven.Client.Lightweight/Connection/ServerClient.cs
	Raven.Client.Lightweight/Document/AbstractDocumentQuery.cs
	Raven.Client.Lightweight/Document/InMemoryDocumentSessionOperations.cs
	Raven.Client.Lightweight/PublicExtensions/LinqExtensions.cs
	Raven.Tests/App.config
	Raven.Tests/AssertAsync.cs
	Raven.Tests/Bugs/Async/Querying.cs
	Raven.Tests/Bugs/IteratingTwice.cs
	Raven.Tests/Bugs/MassivelyMultiTenant.cs
	Raven.Tests/Linq/WhereClause.cs
	Raven.Tests/MailingList/Vlad.cs
	Raven.Tests/Shard/Async/WhenUsingShardedServers.cs
	Raven.Tryouts/app.config
	Utilities/Raven.ProjectRewriter/Program.cs

+++ Raven.Database/Indexing/BaseBatchSizeAutoTuner.cs
@@ -29,7 +29,7 @@ namespace Raven.Database.Indexing
 			}
 		}
 
-		public void AutoThrottleBatchSize(int amountOfItemsToIndex, int size, TimeSpan indexingDuration)
+		public void AutoThrottleBatchSize(int amountOfItemsToIndex, long size, TimeSpan indexingDuration)
 		{
 			try
 			{
@@ -46,7 +46,7 @@ namespace Raven.Database.Indexing
 			}
 		}
 
-		private bool ConsiderIncreasingBatchSize(int amountOfItemsToIndex, int size, TimeSpan indexingDuration)
+		private bool ConsiderIncreasingBatchSize(int amountOfItemsToIndex, long size, TimeSpan indexingDuration)
 		{
 			if (amountOfItemsToIndex < NumberOfItemsToIndexInSingleBatch)
 			{
