+++ Raven.Database/Prefetching/PrefetchingBehavior.cs
@@ -472,18 +472,18 @@ namespace Raven.Database.Prefetching
 				return currentBatchSize / 15;
 			});
 
-			int futureSize = futureIndexBatches.Values.Sum(x =>
+			long futureSize = futureIndexBatches.Values.Sum(x =>
 			{
 				if (x.Task.IsCompleted)
 				{
 					var jsonResults = x.Task.Result;
-					return jsonResults.Sum(s => s.SerializedSizeOnDisk);
+					return jsonResults.Sum(s => (long)s.SerializedSizeOnDisk);
 				}
 				return currentBatchSize * 256;
 			});
 			autoTuner.AutoThrottleBatchSize(
 				jsonDocs.Count + futureLen, 
-				futureSize + jsonDocs.Sum(x => x.SerializedSizeOnDisk),
+				futureSize + jsonDocs.Sum(x => (long)x.SerializedSizeOnDisk),
 			    indexingDuration);
 		}
 	}
